added possibility to list dynamoDB tables to `ls` command
I added the possibility to list dynamo DB tables as well as arbitrary S3 buckets (instead of only bimoage-source-{origin}) because I was exploring the data myself and thought it might come in handy. It can be used as a CLI but I actually added it because the package can be imported and then just use the functions from a python interpreter (instead of every time having to copy paste the code or similar stuff).