Load config.yaml properly
Don't try to work out the path relative to the
executing file - this is prone to errors when
the package is loaded in a way you don't expect
(like installed with pip!).

I moved config.yaml to the biomage directory
so it is included in the package. If people
make config changes in the repo, they need to
re-run `make install`.