Packaging improvements (#54)
* Fix imports

- use relative imports
- add __init__.py to a couple of subpackages
    (this is required for inclusion in the package.
     find_namespace_packages introduces its own problems.)

* Fix installation

- Installation now uses pip, which is far less weird
- The entry point is configured correctly (no need to
   `python3 biomage`)
- The package is now included in the installation so you
   don't have to be in this dir to use the tool
- Reformatted setup.py into logical units
- Split development requirements off from user requirements

* Neaten Makefile echos

* Remove self from requirements.txt

* Install to /usr/bin

Previously, /usr/local/bin was used. This is macOS specific.
The equivalent in most Linux distributions is ~/.local/bin.
I don't believe there is a way to get this path in a cross-
platform way. Also, many (most?) distributions do not even
have the latter on PATH, causing more confusion.

Also adds an uninstall target.

* Fix dev commands

Add new targets to install dev requirements
and modify existing targets to use bins in the
venv. Before this, black and flake8 would have
to have been installed system-wide.

* Remove transitive dependencies from requirements.txt

* Load config.yaml properly

Don't try to work out the path relative to the
executing file - this is prone to errors when
the package is loaded in a way you don't expect
(like installed with pip!).

I moved config.yaml to the biomage directory
so it is included in the package. If people
make config changes in the repo, they need to
re-run `make install`.

* Fix Python 3.7 incompatibility

I believe this is the only code used that is incompatible
with Python 3.7. Several distributions still use Python 3.7
by default, so to prevent having to add a clarification
to the readme, it is easier to just make this change.

* Remove old entry point upon uninstall

* Attempt to fix test workflow

This is unrelated to this PR however I want to
see if this works.

* Install to /usr/local/bin on darwin

* Remove redundant installation step

* Print blank lines under ticks

* Run black

* Reorder imports in `make fmt`

* Reorder imports

* Fix all flake8 warnings

* Stop VSCode from organising imports

* unified sorting & formatting with VS code

* Restore sudo rm